LPS-93977 Refactor the bean-portlet-cdi-extension to use the APIs found in bean-portlet-extension-api and bean-portlet-registration-api
