+++ modules/apps/bean-portlet/bean-portlet-cdi-extension/build.gradle
@@ -9,6 +9,8 @@ dependencies {
 	compileOnly group: "org.osgi", name: "org.osgi.service.http.whiteboard", version: "1.0.0"
 	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
 	compileOnly project(":apps:bean-portlet:bean-portlet-api")
+	compileOnly project(":apps:bean-portlet:bean-portlet-extension-api")
+	compileOnly project(":apps:bean-portlet:bean-portlet-registration-api")
 	compileOnly project(":core:petra:petra-function")
 	compileOnly project(":core:petra:petra-lang")
 	compileOnly project(":core:petra:petra-string")
