+++ modules/apps/bean-portlet/bean-portlet-cdi-extension/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,1 @@
-com.liferay.bean.portlet.cdi.extension.internal.BeanPortletExtension
+com.liferay.bean.portlet.cdi.extension.internal.CDIBeanPortletExtension
